move content from organization lesson
The information from https://datacarpentry.org/organization-geospatial/02-organising-spatial-data/index.html was partially redundant with the information in this episode, but also adds some specific information for working with geospatial data and managing a geospatial project. I've moved that content here and will put in a PR to delete that episode from the organization lesson.