Add item to dplyr enumerated list
In the introduction to the dplyr package the text states that there are 6 useful functions that will be covered. Only 5 are listed in the enumerated list below the text. Towards the end of this module the count() and the n() functions are described but they were not included in the enumerated list.  I added them to the list so that there are now 6 items that are covered.