Lines 123-128
Prior to these lines, the learner has not been taught about the shell environment or completed the shell lessons. The comparison between R and the shell environment is lost on the learner, because they don't have any prior experience (at least, within the lesson) or knowledge of shell environments. I think it would make this section less confusing to remove the shell environment language OR include a brief definition of what that means in order to provide some context, IF keeping the 'shell' language in this section is necessary.

(Note: this proposed change is written in partial fulfillment of my Carpentries instructor course checkout process)- Erika Koontz